Modified island tile creation
Tiles are no longer created when an island is generated, instead they
are placed at a position when they are needed (when foliage is placed,
or wherever a tool will hit (TODO)). Currently places the whole tile
(with all 9 sub-tiles), but Tile's will need to be adjusted to place a
single sub-tile.